pyproject.toml
[project]
name = "sre-agent-system"
version = "0.1.0"
description = "Multi-agent SRE incident response system"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=0.4.0",
    "google-adk>=0.1.0",
    "google-cloud-logging>=3.10.0",
    "google-cloud-compute>=1.19.0",
    "google-cloud-container>=2.45.0",
    "google-auth>=2.29.0",
    "pyyaml>=6.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
